refactor!: rename BiDirectional namespace to TwoWay
