+++ Carbonate/Core/TwoWay/IResponder.cs
@@ -2,7 +2,7 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace Carbonate.Core.BiDirectional;
+namespace Carbonate.Core.TwoWay;
 
 /// <summary>
 /// Gives the ability to respond to a pull request from an <see cref="IReactable{IResponder}"/>.
